Add validation to prevent deletion of last status in category

+++ src/pages/projects/projectView/board/board-section/board-section-card/board-section-card-header.tsx
@@ -177,6 +177,7 @@ const BoardSectionCardHeader: React.FC<BoardSectionCardHeaderProps> = ({
       if (groupBy === IGroupBy.STATUS) {
         const replacingStatusId = '';
         const res = await statusApiService.deleteStatus(groupId, projectId, replacingStatusId);
+        if (res.message === 'At least one status should exists under each category.') return
         if (res.done) {
           dispatch(deleteSection({ sectionId: groupId }));
         } else {
